+++ processing/provider.py
@@ -1,7 +1,6 @@
 # See https://docs.qgis.org/3.10/en/docs/pyqgis_developer_cookbook/processing.html
 from qgis.core import QgsProcessingProvider
 from qgis.PyQt.QtGui import QIcon
-
 from ThreeDiToolbox.processing.threedidepth_algorithm import ThreediDepth
 from ThreeDiToolbox.processing.dwf_calculation_algorithm import DWFCalculatorAlgorithm
 
@@ -21,7 +20,7 @@ class ThreediProvider(QgsProcessingProvider):
         This string should be a unique, short, character only string,
         eg "qgis" or "gdal". This string should not be localised.
         """
-        return 'threedi'
+        return "threedi"
 
     def name(self, *args, **kwargs):
         """The human friendly name of your plugin in Processing.
@@ -29,7 +28,7 @@ class ThreediProvider(QgsProcessingProvider):
         This string should be as short as possible (e.g. "Lastools", not
         "Lastools version 1.0.1 64-bit") and localised.
         """
-        return self.tr('3Di')
+        return self.tr("3Di")
 
     def icon(self):
         """Should return a QIcon which is used for your provider inside