+++ tool_commands/create_breach_locations/command.py
@@ -28,8 +28,8 @@ class CustomCommand(CustomCommandBase):
         self.show_gui()
 
     def show_gui(self):
-        self.tool_dialog_widget = breach_location_dialog.CreateBreachLocationsDialogWidget(
-            command=self
+        self.tool_dialog_widget = (
+            breach_location_dialog.CreateBreachLocationsDialogWidget(command=self)
         )
         self.tool_dialog_widget.exec_()  # block execution