+++ tool_result_selection/result_selection.py
@@ -1,7 +1,5 @@
 # (c) Nelen & Schuurmans, see LICENSE.rst.
 from io import IOBase
-from urllib.parse import urlparse
-
 from qgis.core import QgsNetworkAccessManager
 from qgis.PyQt.QtCore import pyqtSignal
 from qgis.PyQt.QtCore import QObject
@@ -13,6 +11,7 @@ from ThreeDiToolbox.tool_result_selection import models
 from ThreeDiToolbox.tool_result_selection import result_selection_view
 from ThreeDiToolbox.utils.user_messages import messagebar_message
 from ThreeDiToolbox.utils.user_messages import pop_up_info
+from urllib.parse import urlparse
 
 import json
 import logging