+++ utils/user_messages.py
@@ -3,6 +3,7 @@
 from contextlib import contextmanager
 from qgis.core import Qgis
 from qgis.PyQt.QtCore import Qt
+from qgis.PyQt.QtWidgets import QApplication
 from qgis.PyQt.QtWidgets import QMessageBox
 from qgis.PyQt.QtWidgets import QProgressBar
 from qgis.utils import iface
@@ -18,14 +19,14 @@ def pop_up_info(msg="", title="Information", parent=None):
 
 
 def statusbar_message(msg=""):
-    """Display message in status bar """
+    """Display message in status bar"""
     if iface is None:
         return
     iface.mainWindow().statusBar().showMessage(msg)
 
 
 def messagebar_message(title, msg, level=None, duration=0):
-    """ Show message in the message bar (just above the map)
+    """Show message in the message bar (just above the map)
 
     Args:
         title (str): title of messages, showed bold in the start of the message
@@ -67,10 +68,11 @@ class StatusProgressBar(object):
 
         self.message_bar.layout().addWidget(self.progress_bar)
         if iface is not None:
-            iface.messageBar().pushWidget(self.message_bar, Qgis.MessageLevel())
+            iface.messageBar().pushWidget(self.message_bar, Qgis.Info)
 
         self.step_size = 1
         self.progress = 0
+        iface.mainWindow().repaint()
 
     def set_step_size(self, step_size):
 
@@ -82,6 +84,7 @@ class StatusProgressBar(object):
         self.progress_bar.setValue(self.progress)
         if message:
             self.message_bar.setText(message)
+        QApplication.processEvents()
 
     def __del__(self):
         if iface is not None: